+++ tests/testthat/test-fill.R
@@ -103,10 +103,12 @@ test_that("doesn't modify-by-reference", {
 #   expect_equal(attributes(out_u$x), attributes(df$x))
 # })
 
-test_that("fill respects grouping", {
-  df <- data.table::data.table(x = c(1, 1, 2), y = c(1, NA, NA))
-  out <- fill.(df, y, by = x)
-  expect_equal(out$y, c(1, 1, NA))
+test_that("fill respects grouping & is correct order", {
+  df <- data.table::data.table(x = c(1, NA, NA), y = c(1, 1, 2))
+  out <- fill.(df, x, by = y)
+
+  expect_named(out, c("x", "y"))
+  expect_equal(out$x, c(1, 1, NA))
 })
 
 #### Enhanced election