+++ R/tidytable_package.R
@@ -9,4 +9,4 @@ NULL
 
 globalVariables(c("data", ".","..select_vars", ".count", "na_index", ".new_col",
                   "..all_names", "..final_order_i", "..rows", "name", "value",
-                  "..all_cols"))
+                  "..all_cols", "..select_cols", "..keep_names"))