+++ R/arrange.R
@@ -14,7 +14,10 @@
 #'   c = c("a","a","b"))
 #'
 #' example_dt %>%
-#'   arrange.(a, -c)
+#'   arrange.(c, -a)
+#'
+#' example_dt %>%
+#'   arrange.(c, desc.(a))
 arrange. <- function(.data, ...) {
   UseMethod("arrange.")
 }