+++ R/zzz.R
@@ -2,5 +2,12 @@
   requireNamespace("data.table")
   requireNamespace("magrittr")
   requireNamespace("rlang")
+  setHook(packageEvent("knitr", "attach"),
+          function(...) tidytable_first())
   invisible()
+}
+
+tidytable_first <- function(){
+  detach("package:tidytable")
+  library(tidytable, warn.conflicts = FALSE, quietly = TRUE)
 }