+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidytable 0.4.1 (in development)
 
+* Breaking changes:
+  + `get_dummies.()`: "cols" arg defaults to `c(is.character, is.factor)`. The result is the same, but is more consistent with other enhanced selection functions
 * New functions:
   + `desc.()`
   + `unite.()`